+++ src/main/java/com/almostreliable/summoningrituals/recipe/container/RecipeInfo.java
@@ -8,31 +8,33 @@ import net.minecraft.network.codec.StreamCodec;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.item.ItemEntity;
+import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.crafting.RecipeHolder;
 
 import java.util.Collection;
 import java.util.List;
 
 public record RecipeInfo(
-    ResourceLocation getRecipeId, AltarRecipe getRecipe, Collection<Entity> getInputEntities, boolean isBlockPatternExtensionMatched,
-    Collection<ItemEntity> getOutputItems, Collection<Entity> getOutputEntities
+    ResourceLocation getRecipeId, AltarRecipe getRecipe, ItemStack getInitiator, Collection<Entity> getInputEntities,
+    boolean isBlockPatternExtensionMatched, Collection<ItemEntity> getOutputItems, Collection<Entity> getOutputEntities
 ) {
 
     // only used to sync recipe info to custom renderers
     public static final StreamCodec<RegistryFriendlyByteBuf, RecipeInfo> STREAM_CODEC = StreamCodec.composite(
         ResourceLocation.STREAM_CODEC, RecipeInfo::getRecipeId,
         AltarRecipeSerializer.STREAM_CODEC, RecipeInfo::getRecipe,
-        (recipeId, recipe) -> new RecipeInfo(recipeId, recipe, List.of(), false, List.of(), List.of())
+        (recipeId, recipe) -> new RecipeInfo(recipeId, recipe, ItemStack.EMPTY, List.of(), false, List.of(), List.of())
     );
 
-    public static RecipeInfo inputInfo(RecipeHolder<AltarRecipe> recipe, Collection<Entity> inputEntities) {
-        return new RecipeInfo(recipe.id(), recipe.value(), inputEntities, false, List.of(), List.of());
+    public static RecipeInfo inputInfo(RecipeHolder<AltarRecipe> recipe, ItemStack initiator, Collection<Entity> inputEntities) {
+        return new RecipeInfo(recipe.id(), recipe.value(), initiator, inputEntities, false, List.of(), List.of());
     }
 
     public static RecipeInfo blockPatternInfo(RecipeInfo inputInfo, boolean blockPatternExtensionMatched) {
         return new RecipeInfo(
             inputInfo.getRecipeId,
             inputInfo.getRecipe,
+            inputInfo.getInitiator,
             inputInfo.getInputEntities,
             blockPatternExtensionMatched,
             List.of(),
@@ -43,7 +45,7 @@ public record RecipeInfo(
     public static RecipeInfo outputInfo(
         RecipeInfo blockPatternInfo, Collection<ItemEntity> outputItems, Collection<Entity> outputEntities) {
         return new RecipeInfo(
-            blockPatternInfo.getRecipeId, blockPatternInfo.getRecipe, blockPatternInfo.getInputEntities,
+            blockPatternInfo.getRecipeId, blockPatternInfo.getRecipe, blockPatternInfo.getInitiator, blockPatternInfo.getInputEntities,
             blockPatternInfo.isBlockPatternExtensionMatched, outputItems, outputEntities
         );
     }